Создал фикстуру и перенес логику создания экземпляра тестового класса в файл conftest
