+++ proxy/requirements.txt
@@ -1,2 +1,2 @@
 Flask==2.0.2
-
+pytest==6.2.5
